+++ pkg/infrastructure/services/dashboards_test.go
@@ -16,8 +16,14 @@ package services
 
 import (
 	"testing"
+	"time"
 
+	grafanav1 "github.com/integr8ly/grafana-operator/v3/pkg/apis/integreatly/v1alpha1"
+	"github.com/kiegroup/kogito-cloud-operator/pkg/apis/app/v1alpha1"
+	"github.com/kiegroup/kogito-cloud-operator/pkg/test"
 	"github.com/stretchr/testify/assert"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 func Test_fetchDashboardNames(t *testing.T) {
@@ -69,4 +75,49 @@ func Test_fetchDashboards(t *testing.T) {
 	assert.Equal(t, dashboard1, dashboards[0].RawJSONDashboard)
 	assert.Equal(t, dashboard2, dashboards[1].RawJSONDashboard)
 	assert.Equal(t, fetchedDashboardNames[0], dashboards[0].Name)
+}
+
+func Test_serviceDeployer_DeployGrafanaDashboards(t *testing.T) {
+	replicas := int32(1)
+	service := &v1alpha1.KogitoRuntime{
+		ObjectMeta: v1.ObjectMeta{
+			Name:      "my-kogito-runtime",
+			Namespace: t.Name(),
+		},
+		Spec: v1alpha1.KogitoRuntimeSpec{
+			KogitoServiceSpec: v1alpha1.KogitoServiceSpec{Replicas: &replicas},
+		},
+	}
+	cli := test.NewFakeClientBuilder().AddK8sObjects(service).OnOpenShift().Build()
+
+	dashboards := []GrafanaDashboard{
+		{
+			Name:             "mydashboard",
+			RawJSONDashboard: "[]",
+		},
+		{
+			Name:             "myseconddashboard",
+			RawJSONDashboard: "[]",
+		},
+	}
+
+	reconcileAfter, err := deployGrafanaDashboards(dashboards, cli, t.Name())
+	assert.NoError(t, err)
+	assert.Equal(t, time.Duration(0), reconcileAfter)
+
+	dashboard := &grafanav1.GrafanaDashboard{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "mydashboard",
+			Namespace: t.Name(),
+		},
+	}
+	test.AssertFetchMustExist(t, cli, dashboard)
+
+	dashboard = &grafanav1.GrafanaDashboard{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "myseconddashboard",
+			Namespace: t.Name(),
+		},
+	}
+	test.AssertFetchMustExist(t, cli, dashboard)
 }
