+++ pkg/infrastructure/services/deployer.go
@@ -17,14 +17,11 @@ package services
 import (
 	"fmt"
 	"reflect"
-	"strings"
 	"time"
 
 	"github.com/kiegroup/kogito-cloud-operator/pkg/framework"
 	"github.com/kiegroup/kogito-cloud-operator/pkg/infrastructure"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-
-	grafanav1 "github.com/integr8ly/grafana-operator/v3/pkg/apis/integreatly/v1alpha1"
 
 	"github.com/kiegroup/kogito-cloud-operator/pkg/infrastructure/record"
 	v1 "k8s.io/api/core/v1"
@@ -227,43 +224,8 @@ func (s *serviceDeployer) Deploy() (reconcileAfter time.Duration, err error) {
 	}
 
 	reconcileAfter, err = s.configureMessaging()
-	if err != nil || reconcileAfter > 0 {
-		return
-	}
-
-	dashboards, err := FetchGrafanaDashboards(s.client, s.instance)
-	if err != nil {
-		return reconciliationPeriodAfterDashboardsError, err
-	}
-
-	reconcileAfter, err = deployGrafanaDashboards(dashboards, s.client, s.getNamespace())
 
 	return
-}
-
-func deployGrafanaDashboards(dashboards []GrafanaDashboard, cli *client.Client, namespace string) (time.Duration, error) {
-	for _, dashboard := range dashboards {
-		resourceName := strings.ReplaceAll(strings.ToLower(dashboard.Name), ".json", "")
-		dashboardDefinition := &grafanav1.GrafanaDashboard{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      resourceName,
-				Namespace: namespace,
-				Labels: map[string]string{
-					"app": GrafanaDashboardAppName,
-				},
-			},
-			Spec: grafanav1.GrafanaDashboardSpec{
-				Json: dashboard.RawJSONDashboard,
-				Name: dashboard.Name,
-			},
-		}
-		if err := kubernetes.ResourceC(cli).Create(dashboardDefinition); err != nil {
-			log.Warnf("Error occurs while creating dashboard %s, not going to reconcile the resource.", dashboard.Name, err)
-		} else {
-			log.Infof("Successfully created grafana dashboard %s", dashboard.Name)
-		}
-	}
-	return 0, nil
 }
 
 func (s *serviceDeployer) generateEventForDeltaResources(eventReason string, resourceType reflect.Type, addedResources []resource.KubernetesResource) {
@@ -384,5 +346,11 @@ func (s *serviceDeployer) configureMonitoring() (time.Duration, error) {
 	} else if failedVerifyAddon {
 		return reconciliationPeriodMonitoringEndpointNotAvailable, nil
 	}
-	return 0, nil
+
+	reconcileAfter, err := configureGrafanaDashboards(s.client, s.instance, s.getNamespace())
+	if err != nil {
+		return reconciliationPeriodAfterDashboardsError, err
+	}
+
+	return reconcileAfter, err
 }
